Extract Apply actions into their own module
